Edit build.sh

+++ build.sh
@@ -154,15 +154,20 @@ makedist()
 		mkdir -p $DL/deb/dev-$arch
 		mkdir -p $DISTLIB/debian-$arch
 		
-		ar -p $DL/deb/libximc7_*_$arch.deb data.tar.gz | tar zx -C $DL/deb/$arch
-		ar -p $DL/deb/libximc7-dev_*_$arch.deb data.tar.gz | tar zx -C $DL/deb/dev-$arch
+		if [ -f "$DL/deb/libximc7_*_$arch.deb" ]
+		then
+			echo $arch
+			echo $DL/deb/libximc7_*_$arch.deb
+			ar -p $DL/deb/libximc7_*_$arch.deb data.tar.gz | tar zx -C $DL/deb/$arch
+			ar -p $DL/deb/libximc7-dev_*_$arch.deb data.tar.gz | tar zx -C $DL/deb/dev-$arch
 		
-		cp -R $DL/deb/$arch/usr/lib/*.* $DISTLIB/debian-$arch/
-		cp -R $DL/deb/dev-$arch/usr/lib/*.* $DISTLIB/debian-$arch/
-		cp -R $DL/$arch/usr/share/libximc/keyfile.sqlite $DISTLIB/debian-$arch/
+			cp -R $DL/deb/$arch/usr/lib/*.* $DISTLIB/debian-$arch/
+			cp -R $DL/deb/dev-$arch/usr/lib/*.* $DISTLIB/debian-$arch/
+			cp -R $DL/$arch/usr/share/libximc/keyfile.sqlite $DISTLIB/debian-$arch/
+		fi
 		
-		rm -rf $DL/deb/$arch
-		rm -rf $DL/deb/dev-$arch
+		# rm -rf $DL/deb/$arch
+		# rm -rf $DL/deb/dev-$arch
 		
 	done
